scripts/requirements.txt
torch
torchvision
transformers
accelerate
Pillow
numpy
trimesh
rembg
requests
